extract commands and action event into common
