+++ common/src/ipc/mod.rs
@@ -1,0 +1,2 @@
+pub mod action_event;
+pub mod commands;
