+++ common/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod ipc;
 pub mod logging;
