+++ tools/oxide-ipc/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+oxide-common = { version = "0.1.*", path = "../../common" }
+log = "0.4.0"
 zbus = { version = "3", default-features = false, features = ["tokio"] }
 async-std = { version = "1", features = ["attributes", "tokio1"] }
 serde = { version = "1", features = ["derive"] }
