+++ tools/oxide-ipc/src/lib.rs
@@ -1,4 +1,3 @@
-mod events;
 mod ipc;
 pub mod state;
 
@@ -6,8 +5,9 @@ use std::sync::mpsc::Sender;
 use std::sync::{Arc, Mutex};
 
 use crate::ipc::state_signal_channel_async;
-use events::*;
-use ipc::{get_state_async, sent_event_async}; //, wait_for_state_change_async};
+use ipc::{get_state_async, sent_event_async};
+use oxide_common::ipc::action_event::WmActionEvent;
+//, wait_for_state_change_async};
 use state::*;
 
 pub fn get_state() -> String {
